Read a file, filter data, and save for later
A Song data file is read.  The data is filtered for only the wanted
information.  This information is saved into a file.  Another file is
created that can be read quickly and searched in.  This file contains
an off-set of the information in the first file.  The user will input a
song name.  This name is searched for in the second file.  Once the
off-set is found.  The information is gathered from the first file and
printed.