File sorted with individual P-Threads
The starting directory is given with all the unsorted files.  Each file
has a thread created for it.  Qsort is then used to sort the
information based on the database index number.  Once the individual
files are sorted, selected sort is used to put the individual records
to a master, sorted file.