My Own Malloc / Free Functions
My own malloc function uses the SBRK system call to move the heap
pointer and allocate space.  A struct is used for the header data that
holds the pointer to the allocated memory, total size, and it the block
is free or not.  The malloc function will check for previously
allocated blocks that can be re-used instead of move the heap pointer
for each malloc.

When free is called, the block is found by iterated the linked list of
memory blocks.  The previous and next blocks are checked to see if the
blocks can be merged into larger blocks.